PlatformCore\OGLVideoDevice: move shader code into separate header files
